Make KeyframeEffectReadOnly::setKeyframes take effect immediately
Previously we wouldn't notice that the keyframes had changed until the
next animation update. This CL fixes that by clearing any cached sampled
effects and notifying the owning Animation that it needs to invalidate
it's information.

This also fixes updating composited animations; invalidating the effect
on the animation causes us to properly restart the composited animation.

Bug: 799061, 835824
Change-Id: Ib307a5409c52c5dab1ee0f98c989d7dbaef42937